\name{models.payne} 
\docType{data}
\alias{models.payne} 
\title{GammaModel::models.payne} 
\description{
	Probabilities of death in Payne's age classes A to I, 	proposed by Payne to represent the ideal kill strategies 	when attempting to maximise the meat, milk or wool yields 	from a caprine herd. 
	}
\usage{Neolithic}
\format{
	A data frame comprising 27 probabilities, across 3 rows and 9 columns.
	}
\source{
	Data as numeric probabilities was first published by 	Timpson et al 2018. However, this numeric summary was 	derived from the original source data resented 	graphically as survivorship curves in Table 1 of Payne, 	S., 1973. 
	}
\references{
	Modelling caprine age-at-death profiles using the Gamma 	distribution, Adrian Timpson, Rosalind E. Gillis, 	Katie Manning and Mark G. Thomas, Journal of 	Archaeological Science 2018

	Payne, S., 1973. Kill-off patterns in sheep and goats: 	the 	mandibles from Asvan Kale, Anatolian studies.
	}