\name{models.redding} 
\docType{data}
\alias{models.redding} 
\title{GammaModel::models.redding} 
\description{
	Probabilities of death in Payne's age classes A to I, 	proposed by Redding to represent the ideal kill strategies 	when attempting to maximise energy yields or herd security 	from a caprine herd. 
	}
\usage{Neolithic}
\format{
	A data frame comprising 16 probabilities, across 2 rows 	and 8 columns.
	}
\source{
	Data as numeric probabilities was first published in 	Redding, R.,1981. 
	}
\references{
	Redding, R.,1981. Decision making in subsistence herding 	of sheep and goats in the Middle East. University of 	Michigan, Ann Arbor. 	Ph. D.
	}