\name{Neolithic} 
\docType{data}
\alias{Neolithic} 
\title{GammaModel::Neolithic} 
\description{
	Age-at-death data comprising counts of caprine teeth from 	10 European Neolithic sites, classified into Payne's age 	classes A to I. Some individual teeth have been assigned 	to several possible age classes (a multi-class). 
	}
\usage{Neolithic}
\format{
	A data frame comprising 719 counts, across 10 rows and 	18 columns
	}
\source{
	Original source data first published in Timpson et al 	2018.
	}
\references{
	Modelling caprine age-at-death profiles using the Gamma 	distribution. Adrian Timpson, Rosalind E. Gillis, 	Katie Manning and Mark G. Thomas. Journal of 	Archaeological Science 2018
	}